# docs/swagger.json
{
    "schemes": [
        "HTTP"
    ],
    "swagger": "2.0",
    "info": {
        "description": "下一代媒体刮削\u0026整理工具",
        "title": "MediaTools API 文档",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/config/fanart": {
            "get": {
                "description": "获取 Fanart 配置",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "获取 Fanart 配置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.FanartConfig"
                        }
                    }
                }
            },
            "post": {
                "description": "更新 Fanart 配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "更新 Fanart 配置",
                "parameters": [
                    {
                        "description": "Fanart 配置",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/config.FanartConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.FanartConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrResponse"
                        }
                    }
                }
            }
        },
        "/config/libraries": {
            "get": {
                "description": "获取媒体库配置",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "获取媒体库配置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/config.LibraryConfig"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "更新媒体库配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "更新媒体库配置",
                "parameters": [
                    {
                        "description": "媒体库配置",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/config.LibraryConfig"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/config.LibraryConfig"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrResponse"
                        }
                    }
                }
            }
        },
        "/config/log": {
            "get": {
                "description": "获取日志配置",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "获取日志配置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.LogConfig"
                        }
                    }
                }
            },
            "post": {
                "description": "更新日志配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "更新日志配置",
                "parameters": [
                    {
                        "description": "日志配置",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/config.LogConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.LogConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrResponse"
                        }
                    }
                }
            }
        },
        "/config/media/custom_word": {
            "get": {
                "description": "获取自定义词配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "获取自定义词配置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.CustomWordConfig"
                        }
                    }
                }
            },
            "post": {
                "description": "更新自定义词配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "更新自定义词配置",
                "parameters": [
                    {
                        "description": "自定义词配置",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/config.CustomWordConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.CustomWordConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrResponse"
                        }
                    }
                }
            }
        },
        "/config/media/format": {
            "get": {
                "description": "获取媒体格式配置",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "获取媒体格式配置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.FormatConfig"
                        }
                    }
                }
            },
            "post": {
                "description": "更新媒体格式配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "更新媒体格式配置",
                "parameters": [
                    {
                        "description": "媒体格式配置",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/config.FormatConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.FormatConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrResponse"
                        }
                    }
                }
            }
        },
        "/config/tmdb": {
            "get": {
                "description": "获取 TMDB 配置",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "获取 TMDB 配置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.TMDBConfig"
                        }
                    }
                }
            },
            "post": {
                "description": "更新 TMDB 配置",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "更新 TMDB 配置",
                "parameters": [
                    {
                        "description": "TMDB 配置",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/config.TMDBConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.TMDBConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "config.CustomWordConfig": {
            "type": "object",
            "properties": {
                "customization": {
                    "description": "自定义占位置词",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "exclude_words": {
                    "description": "自定义排除词",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "identify_word": {
                    "description": "自定义识别词",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "config.FanartConfig": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "api_url": {
                    "type": "string"
                },
                "languages": {
                    "description": "语言顺序",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "config.FormatConfig": {
            "type": "object",
            "properties": {
                "movie": {
                    "description": "电影格式",
                    "type": "string"
                },
                "tv": {
                    "description": "电视剧格式",
                    "type": "string"
                }
            }
        },
        "config.LibraryConfig": {
            "type": "object",
            "properties": {
                "dst_path": {
                    "description": "目标路径",
                    "type": "string"
                },
                "dst_type": {
                    "description": "目标类型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schemas.StorageType"
                        }
                    ]
                },
                "name": {
                    "description": "媒体库名称",
                    "type": "string"
                },
                "organize_by_category": {
                    "description": "是否按分类分文件夹",
                    "type": "boolean"
                },
                "organize_by_type": {
                    "description": "是否按类型分文件夹",
                    "type": "boolean"
                },
                "src_path": {
                    "description": "源路径",
                    "type": "string"
                },
                "src_type": {
                    "description": "源类型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schemas.StorageType"
                        }
                    ]
                },
                "transfer_type": {
                    "description": "传输类型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schemas.TransferType"
                        }
                    ]
                }
            }
        },
        "config.LogConfig": {
            "type": "object",
            "properties": {
                "console_level": {
                    "description": "终端输出日志级别",
                    "type": "string"
                },
                "file_dir": {
                    "description": "日志文件目录",
                    "type": "string"
                },
                "file_level": {
                    "description": "文件日志级别",
                    "type": "string"
                }
            }
        },
        "config.TMDBConfig": {
            "type": "object",
            "properties": {
                "api_key": {
                    "description": "API Key",
                    "type": "string"
                },
                "api_url": {
                    "description": "TMDB API URL",
                    "type": "string"
                },
                "image_url": {
                    "description": "图片 API URL",
                    "type": "string"
                },
                "include_image_language": {
                    "description": "包含的图片语言",
                    "type": "string"
                },
                "language": {
                    "description": "语言",
                    "type": "string"
                }
            }
        },
...
        "schemas.ErrResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "schemas.StorageType": {
            "type": "integer",
            "format": "int32",
            "enum": [
                0,
                1
            ],
            "x-enum-comments": {
                "StorageLocal": "本地文件系统",
                "StorageUnknown": "未知文件系统"
            },
            "x-enum-descriptions": [
                "未知文件系统",
                "本地文件系统"
            ],
            "x-enum-varnames": [
                "StorageUnknown",
                "StorageLocal"
            ]
        },
        "schemas.TransferType": {
            "type": "integer",
            "format": "int32",
            "enum": [
                0,
                1,
                2,
                3,
                4
            ],
            "x-enum-comments": {
                "TransferCopy": "复制",
                "TransferLink": "硬链接",
                "TransferMove": "移动",
                "TransferSoftLink": "软链接",
                "TransferUnknown": "未知传输类型"
            },
            "x-enum-descriptions": [
                "未知传输类型",
                "复制",
                "移动",
                "硬链接",
                "软链接"
            ],
            "x-enum-varnames": [
                "TransferUnknown",
                "TransferCopy",
                "TransferMove",
                "TransferLink",
                "TransferSoftLink"
            ]
        }
    }
}