# internal/router/config/media.go
package config

import (
	"MediaTools/internal/config"
	"MediaTools/internal/controller/media_controller"
	"MediaTools/internal/schemas"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/sirupsen/logrus"
)

// @Router /config/libraries [get]
// @Summary 获取媒体库配置
// @Description 获取媒体库配置
// @Tags config
// @Produce json
// @Success 200 {object} []config.LibraryConfig
func MediaLibrary(ctx *gin.Context) {
	ctx.JSON(http.StatusOK, config.Media.Libraries)
}

// @Router /config/libraries [post]
// @Summary 更新媒体库配置
// @Description 更新媒体库配置
// @Tags config
// @Accept json
// @Produce json
// @Param config body []config.LibraryConfig true "媒体库配置"
// @Success 200 {object} []config.LibraryConfig
// @Failure 400 {object} schemas.ErrResponse
// @Failure 500 {object} schemas.ErrResponse
func UpdateMediaLibrary(ctx *gin.Context) {
	var (
		req     []config.LibraryConfig
		errResp schemas.ErrResponse
	)

	err := ctx.ShouldBindJSON(&req)
	if err != nil {
		errResp.Message = "请求参数错误: " + err.Error()
		ctx.JSON(http.StatusBadRequest, errResp)
		return
	}

	logrus.Debugf("开始更新媒体库配置: %+v", req)

	oldConfig := config.Media.Libraries
	config.Media.Libraries = req
	err = media_controller.Init()
	if err != nil {
		logrus.Errorf("初始化 Media 控制器失败: %v", err)
		errResp.Message = "初始化 Media 控制器失败: " + err.Error()
		ctx.JSON(http.StatusInternalServerError, errResp)
		goto initErr
	}

	logrus.Debugf("Media 控制器初始化成功: %+v", config.Media.Libraries)

	err = config.WriteConfig()
	if err != nil {
		errResp.Message = "更新配置失败: " + err.Error()
		ctx.JSON(http.StatusInternalServerError, errResp)
		return
	}

	ctx.JSON(http.StatusOK, config.Media.Libraries)
	return
initErr:
	config.Media.Libraries = oldConfig
	media_controller.Init()
}

// @Router /config/media/format [get]
// @Summary 获取媒体格式配置
// @Description 获取媒体格式配置
// @Tags config
// @Produce json
// @Success 200 {object} config.FormatConfig
func MediaFormat(ctx *gin.Context) {
	ctx.JSON(http.StatusOK, config.Media.Format)
}

// @Router /config/media/format [post]
// @Summary 更新媒体格式配置
// @Description 更新媒体格式配置
// @Tags config
// @Accept json
// @Produce json
// @Param config body config.FormatConfig true "媒体格式配置"
// @Success 200 {object} config.FormatConfig
// @Failure 400 {object} schemas.ErrResponse
// @Failure 500 {object} schemas.ErrResponse
func UpdateMediaFormat(ctx *gin.Context) {
	var (
		req     config.FormatConfig
		ErrResp schemas.ErrResponse
	)

	err := ctx.ShouldBindJSON(&req)
	if err != nil {
		ErrResp.Message = "请求参数错误: " + err.Error()
		ctx.JSON(http.StatusBadRequest, ErrResp)
		return
	}

	oldConfig := config.Media.Format
	config.Media.Format = req
	err = media_controller.InitFormatTemplates()
	if err != nil {
		ErrResp.Message = "初始化格式模板失败: " + err.Error()
		ctx.JSON(http.StatusInternalServerError, ErrResp)
		goto initErr
	}
	err = config.WriteConfig()
	if err != nil {
		ErrResp.Message = "更新配置失败: " + err.Error()
		ctx.JSON(http.StatusInternalServerError, ErrResp)
		return
	}

	ctx.JSON(http.StatusOK, config.Media.Format)
	return
initErr:
	config.Media.Format = oldConfig
	media_controller.Init()
}

// @Router /config/media/custom_word [get]
// @Summary 获取自定义词配置
// @Description 获取自定义词配置
// @Tags config
// @Accept json
// @Produce json
// @Success 200 {object} config.CustomWordConfig
func CustomWord(ctx *gin.Context) {
	ctx.JSON(http.StatusOK, config.Media.CustomWord)
}

// @Router /config/media/custom_word [post]
// @Summary 更新自定义词配置
// @Description 更新自定义词配置
// @Tags config
// @Accept json
// @Produce json
// @Param config body config.CustomWordConfig true "自定义词配置"
// @Success 200 {object} config.CustomWordConfig
// @Failure 400 {object} schemas.ErrResponse
// @Failure 500 {object} schemas.ErrResponse
func UpdateCustomWord(ctx *gin.Context) {
	var (
		req     config.CustomWordConfig
		errResp schemas.ErrResponse
	)

	err := ctx.ShouldBindJSON(&req)
	if err != nil {
		errResp.Message = "请求参数错误: " + err.Error()
		ctx.JSON(http.StatusBadRequest, errResp)
		return
	}

	oldConfig := config.Media.CustomWord
	config.Media.CustomWord = req
	err = media_controller.InitCustomWord()
	if err != nil {
		errResp.Message = "初始化自定义词失败: " + err.Error()
		ctx.JSON(http.StatusInternalServerError, errResp)
		goto initErr
	}

	err = config.WriteConfig()
	if err != nil {
		errResp.Message = "更新配置失败: " + err.Error()
		ctx.JSON(http.StatusInternalServerError, errResp)
		return
	}

	ctx.JSON(http.StatusOK, config.Media.CustomWord)
	return
initErr:
	config.Media.CustomWord = oldConfig
	media_controller.Init()
}